# specs/1-rag-chatbot/plan/contracts/api-contract.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API for Docusaurus Book
  description: API for the RAG chatbot system embedded in Docusaurus book
  version: 1.0.0
servers:
  - url: https://api.rag-chatbot.example.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/documents/ingest:
    post:
      summary: Ingest documents for RAG
      description: Processes and ingests book documents for RAG functionality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - source_path
                - content
              properties:
                source_path:
                  type: string
                  description: Path to the source document
                  example: "/docs/intro.md"
                content:
                  type: string
                  description: Document content to be processed
                metadata:
                  type: object
                  description: Additional metadata for the document
      responses:
        '200':
          description: Documents ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: Processing status
                    example: "success"
                  processed_chunks:
                    type: integer
                    description: Number of content chunks processed
                    example: 15
                  document_id:
                    type: string
                    description: ID of the processed document
        '400':
          description: Invalid document format
        '500':
          description: Internal server error
...